some updates

+++ RL/main.py
@@ -8,9 +8,13 @@ os.environ["CUDA_VISIBLE_DEVICES"] = "1"
 
 if __name__ == '__main__':
     root_dir = VARIANT['log_path']
-    os.makedirs(os.path.dirname(root_dir + "/config_out.json"), exist_ok=True)
-    with open(root_dir + "/config_out.json", 'w') as fp:
-        json.dump(VARIANT, fp)
+    file_path = root_dir + "/config_out.json"
+    if os.path.isfile(file_path):
+        pass
+    else:
+        os.makedirs(os.path.dirname(file_path), exist_ok=True)
+        with open(file_path, 'w') as fp:
+            json.dump(VARIANT, fp)
     if VARIANT['train']:
         for i in range(VARIANT['start_of_trial'], VARIANT['start_of_trial']+VARIANT['num_of_trials']):
             VARIANT['log_path'] = root_dir +'/'+ str(i)
